# Speed up result counting