**What**:

Speed up result counting iterator.

2 parts:
1. Use result_new_severities_dynamic/result_new_severities_static instead of result_new_severities when the dynamic state is known.
2. Use LATERAL to save having the severity calculation repeated in the WHERE clause (where it was not accessing the WITH optimisations).

Test case was an `--optimize rebuild-report-cache` that took 32s on master.

With 1 the test case dropped to 23s, with 1 and 2 it dropped to under 4s.

<!--
  Describe what changes are being made, e.g. which feature/bug is being
  developed/fixed in this PR?
-->

**Why**:

Immediate problem: Result counting slows down the report upload used to generate big dbs.

<!-- Why are these changes necessary? -->

**How did you test it**:

1. `time gvmd --optimize rebuild-report-cache` on master (32s)
2. `psql gvmd -c "SELECT report, user, severity, count, override, end_time, min_qod from report_counts;" > /tmp/rc_master`
3. `time gvmd --optimize rebuild-report-cache` with patch (4s)
4. `psql gvmd -c "SELECT report, user, severity, count, override, end_time, min_qod from report_counts;" | sort > /tmp/rc_pr`
5. `sort /tmp/rc_master | diff - /tmp/rc_pr` (same)

<!--
  How did you verify the changes in this PR?
  If this PR contains tests this section can be considered done.
  Otherwise please write down the steps on how you did test your changes and
  verified that the changes are working as expected.

  See https://www.ministryoftesting.com/dojo/lessons/community-stories-to-shift-left-start-right
  for some background.
 -->

**Checklist**:

<!-- add "N/A" to the end of each line not applicable to your changes -->

<!-- to check an item, place an "x" in the box like so: "- [x] Tests" -->

- [ ] Tests
- [x] [CHANGELOG](https://github.com/greenbone/gvmd/blob/master/CHANGELOG.md) Entry